refactor: remove some unused code and break some circular dependencies
