+++ frontend/app/src/composables/session/use-items-per-page.ts
@@ -1,0 +1,7 @@
+/**
+ * Keeps track of a shared, global instance of the items per page setting.
+ *
+ * It is shared between the main.ts and settings store, because referencing the store
+ * directly, creates issues with tracking.
+ */
+export const useItemsPerPage = createSharedComposable(() => ref<number>(10));
