+++ frontend/app/src/composables/use-pagination-filter/index.ts
@@ -6,8 +6,8 @@ import { defaultCollectionState } from '@/utils/collection';
 import { logger } from '@/utils/logging';
 import { nonEmptyProperties } from '@/utils/data';
 import { useNotificationsStore } from '@/store/notifications';
-import { useItemsPerPage } from '@/composables/session/settings';
 import { applyPaginationDefaults, applySortingDefaults, getApiSortingParams, parseQueryHistory, parseQueryPagination } from '@/composables/use-pagination-filter/utils';
+import { useItemsPerPage } from '@/composables/session/use-items-per-page';
 import type { LocationQuery, RawLocationQuery } from '@/types/route';
 import type { DataTableSortData, TablePaginationData } from '@rotki/ui-library';
 import type { MaybeRef } from '@vueuse/core';
