+++ frontend/app/src/composables/user/use-logged-user-identifier.ts
@@ -1,0 +1,1 @@
+export const useLoggedUserIdentifier = createSharedComposable(() => useSessionStorage<string | undefined>('rotki.logged_user_id', undefined));
