+++ frontend/app/src/locales/cn.json
@@ -2515,13 +2515,7 @@
     }
   },
   "kraken_staking_events": {
-    "refresh_tooltip": "刷新您的 Kraken 质押事件",
-    "types": {
-      "receive_staked_asset": "接收质押资产",
-      "stake_asset": "质押资产",
-      "staking_reward": "质押奖励",
-      "unstake_asset": "取消抵押资产"
-    }
+    "refresh_tooltip": "刷新您的 Kraken 质押事件"
   },
   "kraken_staking_overview": {
     "current": "当前",
