+++ frontend/app/src/store/blockchain/accounts/migrate.ts
@@ -1,12 +1,12 @@
 import { type Notification, NotificationCategory, Severity } from '@rotki/common';
 import { type MaybeRef, useSessionStorage } from '@vueuse/core';
 import { startPromise } from '@shared/utils';
-import { useLoggedUserIdentifier } from '@/composables/user/account';
 import { useSessionAuthStore } from '@/store/session/auth';
 import { useNotificationsStore } from '@/store/notifications';
 import { useBlockchains } from '@/composables/blockchain';
 import { useTokenDetection } from '@/composables/balances/token-detection';
 import { useSupportedChains } from '@/composables/info/chains';
+import { useLoggedUserIdentifier } from '@/composables/user/use-logged-user-identifier';
 import type { MigratedAddresses } from '@/types/websocket-messages';
 
 function setupMigrationSessionCache(identifier: string): Ref<MigratedAddresses> {
