+++ frontend/app/src/store/settings/frontend.ts
@@ -8,8 +8,8 @@ import {
 import { logger } from '@/utils/logging';
 import { useSettingsApi } from '@/composables/api/settings/settings-api';
 import { useLastLanguage } from '@/composables/session/language';
-import { useItemsPerPage } from '@/composables/session/settings';
 import { useComputedRef } from '@/composables/utils/useComputedRef';
+import { useItemsPerPage } from '@/composables/session/use-items-per-page';
 import type { ActionStatus } from '@/types/action';
 
 export const useFrontendSettingsStore = defineStore('settings/frontend', () => {
