+++ frontend/app/src/types/filtering.ts
@@ -1,10 +1,5 @@
 import { z } from 'zod';
-import { convertFromTimestamp, convertToTimestamp } from '@/utils/date';
 import { AssetInfoWithId, type AssetsWithId } from '@/types/asset';
-import type { AssetInfo } from '@rotki/common';
-import type { DateFormat } from '@/types/date-format';
-import type { AssetSearchParams } from '@/composables/api/assets/info';
-import type { ComputedRef, Ref } from 'vue';
 
 export enum FilterBehaviour {
   INCLUDE = 'include',
@@ -83,60 +78,4 @@ export enum SavedFilterLocation {
   HISTORY_EVENTS = 'historyEvents',
   BLOCKCHAIN_ACCOUNTS = 'blockchainAccounts',
   ETH_VALIDATORS = 'ethValidators',
-}
-
-export function assetSuggestions(assetSearch: (params: AssetSearchParams) => Promise<AssetsWithId>, evmChain?: string): (value: string) => Promise<AssetsWithId> {
-  let pending: AbortController | null = null;
-
-  return useDebounceFn(async (value: string) => {
-    if (pending) {
-      pending.abort();
-      pending = null;
-    }
-
-    pending = new AbortController();
-
-    let keyword = value;
-    let address;
-
-    if (isValidEthAddress(value)) {
-      keyword = '';
-      address = value;
-    }
-
-    const result = await assetSearch({
-      address,
-      evmChain,
-      limit: 10,
-      signal: pending.signal,
-      value: keyword,
-    });
-    pending = null;
-    return result;
-  }, 200);
-}
-
-export function assetDeserializer(assetInfo: (identifier: string) => ComputedRef<AssetInfo | null>): (identifier: string) => AssetInfoWithId | null {
-  return (identifier: string): AssetInfoWithId | null => {
-    const asset = get(assetInfo(identifier));
-    if (!asset)
-      return null;
-
-    return {
-      ...asset,
-      identifier,
-    };
-  };
-}
-
-export function dateValidator(dateInputFormat: Ref<DateFormat>): (value: string) => boolean {
-  return (value: string) => value.length > 0 && !isNaN(convertToTimestamp(value, get(dateInputFormat)));
-}
-
-export function dateSerializer(dateInputFormat: Ref<DateFormat>): (date: string) => string {
-  return (date: string) => convertToTimestamp(date, get(dateInputFormat)).toString();
-}
-
-export function dateDeserializer(dateInputFormat: Ref<DateFormat>): (timestamp: string) => string {
-  return (timestamp: string) => convertFromTimestamp(parseInt(timestamp), get(dateInputFormat));
 }
