+++ frontend/app/src/utils/blockchain/accounts/create-account-with-balance.ts
@@ -1,0 +1,20 @@
+import { getAccountBalance } from '@/utils/blockchain/accounts/index';
+import { getAccountAddress } from '@/utils/blockchain/accounts/utils';
+import type { BlockchainAccount, BlockchainAccountWithBalance } from '@/types/blockchain/accounts';
+import type { BlockchainAssetBalances } from '@/types/blockchain/balances';
+
+export function createAccountWithBalance(
+  account: BlockchainAccount,
+  chainBalances: BlockchainAssetBalances,
+): BlockchainAccountWithBalance {
+  const { balance, expansion } = getAccountBalance(account, chainBalances);
+  const address = getAccountAddress(account);
+
+  return {
+    groupId: address,
+    type: 'account',
+    ...account,
+    ...balance,
+    expansion,
+  } satisfies BlockchainAccountWithBalance;
+}
